Simplify the PR to limit changes to unrelated components
